move run() and build_app() to startup.rs; move request handler to routes (for now)
